refactor(web): move species catalog API into species_catalog pkg
api.py + shim species_catalog_api_service. Refs #344 phase B.

Made-with: Cursor
